+++ tools/diagraphica/cluster.h
@@ -23,13 +23,13 @@ class Cluster
 public:
 	// -- constructors and destructors ------------------------------
     Cluster();
-    Cluster( const std::vector< int > &crd );
+    Cluster( const std::vector< size_t > &crd );
     Cluster( const Cluster &clst );
     virtual ~Cluster();
 
     // -- set functions ---------------------------------------------
-    void setCoord( const std::vector< int > &crd  );
-    void setIndex( const int &idx );
+    void setCoord( const std::vector< size_t > &crd  );
+    void setIndex( const size_t &idx );
     void setParent( Cluster* p );
     void addChild( Cluster* c );
     void setChildren( const std::vector< Cluster* > &c );
@@ -37,7 +37,7 @@ public:
     void setNodes( const std::vector< Node* > &n );
 
     void setAttribute( Attribute* attr );
-    void setAttrValIdx( const int &idx );
+    void setAttrValIdx( const size_t &idx );
 
     void addInBundle( Bundle* b );
     void setInBundles( const std::vector< Bundle* > b );
@@ -45,24 +45,24 @@ public:
     void setOutBundles( const std::vector< Bundle* > b );
 
     // -- get functions ---------------------------------------------
-    int getSizeCoord();
-    int getCoord( const int &idx );
-    void getCoord( std::vector< int > &crd );
-    int getIndex();
+    size_t getSizeCoord();
+    size_t getCoord( const size_t &idx );
+    void getCoord( std::vector< size_t > &crd );
+    size_t getIndex();
     Cluster* getParent();
-    int getSizeChildren();
-    Cluster* getChild( const int &idx );
-    int getSizeNodes();
-    Node* getNode( const int &idx );
-    int getSizeDescNodes();
+    size_t getSizeChildren();
+    Cluster* getChild( const size_t &idx );
+    size_t getSizeNodes();
+    Node* getNode( const size_t &idx );
+    size_t getSizeDescNodes();
 
     Attribute* getAttribute();
-    int getAttrValIdx();
+    size_t getAttrValIdx();
 
-    int getSizeInBundles();
-    Bundle* getInBundle( const int &idx );
-    int getSizeOutBundles();
-    Bundle* getOutBundle( const int &idx );
+    size_t getSizeInBundles();
+    Bundle* getInBundle( const size_t &idx );
+    size_t getSizeOutBundles();
+    Bundle* getOutBundle( const size_t &idx );
 
     // -- clear functions -------------------------------------------
     void clearParent();
@@ -74,16 +74,16 @@ public:
 
 protected:
     // -- utility functions -----------------------------------------
-    void getSizeDescNodes( Cluster* curClst, int &sum );
+    void getSizeDescNodes( Cluster* curClst, size_t &sum );
 
     // -- data members ----------------------------------------------
-    std::vector< int >      coord;
-    int                index;
+    std::vector< size_t >      coord;
+    size_t                index;
     Cluster*           parent;    // association
     std::vector< Cluster* > children;  // association
     std::vector< Node* >    nodes;     // association
     Attribute*         attribute; // association
-    int                attrValIdx;
+    size_t                attrValIdx;
     std::vector< Bundle* >  inBundles;  // association
     std::vector< Bundle* >  outBundles; // association
 };
